+++ CHANGELOG.md
@@ -20,3 +20,31 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Security
 
 [Unreleased]: https://github.com/OXID-eSales/paymorrow-module/compare/HEAD...HEAD
+
+## [v2.0.1]
+
+### Changed
+ - Improved payment data workflows in checkout process
+
+## [v2.0.0]
+
+### Added
+ - Compatibility with OXID eShop v6.x
+ 
+### Changed
+ - Acceptance and unit tests adjusted to run on OXID Testing Library v2.1.0+
+ - All files encoding changed to UTF-8
+ 
+### Fixed
+ - Payment ID handling and Wrapping Cost VAT fixes for the eShop v6
+
+## [v1.0.1]
+
+### Added
+ - User manuals extended for information on module integration in highly customized shops
+
+### Changed
+ - Paymorrow gateway updated to version "patch_20150129"
+ 
+## [v1.0.0]
+Initial release.